embeddings.py
"""
Community Detection and Embedding Module
Uses Gemini LLM for real semantic embeddings with TF-IDF fallback.
"""
import numpy as np
import networkx as nx
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.decomposition import PCA
import random, os, time

random.seed(42)
np.random.seed(42)


# ─────────────────────────────────────────────────────────────────────────────
# Community Detection
# ─────────────────────────────────────────────────────────────────────────────

def detect_communities_louvain(G):
    if G.number_of_edges() == 0:
        return {node: i for i, node in enumerate(G.nodes())}
    communities = nx.community.greedy_modularity_communities(G)
    return {node: i for i, comm in enumerate(communities) for node in comm}


def detect_communities_label_propagation(G):
    if G.number_of_edges() == 0:
        return {node: i for i, node in enumerate(G.nodes())}
    communities = nx.community.label_propagation_communities(G)
    return {node: i for i, comm in enumerate(communities) for node in comm}


def compute_modularity(G, node_community):
    communities = {}
    for node, comm in node_community.items():
        communities.setdefault(comm, set()).add(node)
    try:
        return nx.community.modularity(G, list(communities.values()))
    except Exception:
        return 0.0


# ─────────────────────────────────────────────────────────────────────────────
# Gemini LLM Embeddings (Real) with TF-IDF fallback
# ─────────────────────────────────────────────────────────────────────────────

class GeminiEmbedder:
    """
    Real LLM embeddings via Google Gemini text-embedding-004.
    Falls back to TF-IDF + PCA if API unavailable.
    """

    EMBEDDING_DIM = 768  # Gemini text-embedding-004

    def __init__(self, api_key=None, n_components=32):
        self.api_key = api_key or os.getenv("GEMINI_API_KEY") or os.getenv("GOOGLE_API_KEY")
        self.n_components = n_components
        self.use_llm = False
        self.genai = None
        self._init_client()

    def _init_client(self):
        try:
            import google.generativeai as genai
            if self.api_key:
                genai.configure(api_key=self.api_key)
                self.genai = genai
                self.use_llm = True
        except ImportError:
            self.use_llm = False

    def _embed_gemini(self, texts):
        """Call Gemini API, return (N, 768) ndarray."""
        embeddings = []
        for i, text in enumerate(texts):
            try:
                result = self.genai.embed_content(
                    model="models/text-embedding-004",
                    content=text[:2000],  # truncate to safe limit
                    task_type="SEMANTIC_SIMILARITY"
                )
                embeddings.append(result['embedding'])
                if (i + 1) % 15 == 0:
                    time.sleep(1.0)   # respect quota
            except Exception:
                embeddings.append(np.random.randn(self.EMBEDDING_DIM).tolist())
        return np.array(embeddings, dtype=float)

    def _tfidf_pca(self, texts):
        """TF-IDF + PCA fallback — returns (N, n_components)."""
        vectorizer = TfidfVectorizer(max_features=500, ngram_range=(1, 2), sublinear_tf=True)
        tfidf = vectorizer.fit_transform(texts).toarray()
        n_comp = min(self.n_components, tfidf.shape[0], tfidf.shape[1])
        pca = PCA(n_components=n_comp, random_state=42)
        return pca.fit_transform(tfidf)

    def fit_transform(self, texts):
        """Embed texts and reduce to n_components via PCA."""
        if self.use_llm:
            try:
                raw = self._embed_gemini(texts)
                n_comp = min(self.n_components, raw.shape[0], raw.shape[1])
                pca = PCA(n_components=n_comp, random_state=42)
                return pca.fit_transform(raw)
            except Exception:
                pass
        return self._tfidf_pca(texts)


# Backward-compatible alias
LLMEmbedder = GeminiEmbedder


def compute_semantic_embeddings(cvs, jobs):
    """Compute Gemini LLM embeddings for all CVs and Jobs."""
    all_texts = [cv["text"] for cv in cvs] + [job["text"] for job in jobs]
    embedder = GeminiEmbedder(n_components=32)
    all_embeddings = embedder.fit_transform(all_texts)
    cv_embeddings  = {cv["id"]:  all_embeddings[i]             for i, cv  in enumerate(cvs)}
    job_embeddings = {job["id"]: all_embeddings[len(cvs) + j]  for j, job in enumerate(jobs)}
    return cv_embeddings, job_embeddings, embedder


def predict_links_llm(cv_embeddings, job_embeddings, G, top_k=5, threshold=-0.1):
    """Predict missing CV-Job links via cosine similarity of LLM embeddings."""
    cv_ids = list(cv_embeddings.keys())
    job_ids = list(job_embeddings.keys())
    cv_matrix  = np.array([cv_embeddings[cid]  for cid  in cv_ids])
    job_matrix = np.array([job_embeddings[jid] for jid in job_ids])
    sim_matrix = cosine_similarity(cv_matrix, job_matrix)

    predictions = []
    for i, cv_id in enumerate(cv_ids):
        for j, job_id in enumerate(job_ids):
            if not G.has_edge(cv_id, job_id):
                score = float(sim_matrix[i, j])
                if score > threshold:
                    predictions.append((cv_id, job_id, score))
    predictions.sort(key=lambda x: -x[2])
    return predictions


# ─────────────────────────────────────────────────────────────────────────────
# Node Features
# ─────────────────────────────────────────────────────────────────────────────

def build_node_features(cvs, jobs, G, cv_embeddings, job_embeddings, node_community):
    """Combine structural + semantic + community features for classification."""
    features = {}
    degree_cent = nx.degree_centrality(G)
    betweenness  = nx.betweenness_centrality(G, normalized=True)
    closeness    = nx.closeness_centrality(G)
    emb_dim = len(next(iter(cv_embeddings.values()))) if cv_embeddings else 32

    for cv in cvs:
        nid = cv["id"]
        structural = [G.degree(nid), degree_cent.get(nid,0), betweenness.get(nid,0),
                      closeness.get(nid,0), cv["years_exp"], len(cv["skills"]), int(cv["polyvalent"])]
        semantic = cv_embeddings.get(nid, np.zeros(emb_dim)).tolist()
        features[nid] = structural + semantic + [node_community.get(nid, -1)]

    for job in jobs:
        nid = job["id"]
        structural = [G.degree(nid), degree_cent.get(nid,0), betweenness.get(nid,0),
                      closeness.get(nid,0), 0, len(job["skills_required"]), 0]
        semantic = job_embeddings.get(nid, np.zeros(emb_dim)).tolist()
        features[nid] = structural + semantic + [node_community.get(nid, -1)]

    return features


def compute_2d_projection(features_dict):
    node_ids = list(features_dict.keys())
    X = np.array([features_dict[n] for n in node_ids])
    pca = PCA(n_components=2, random_state=42)
    X2d = pca.fit_transform(X)
    return {nid: X2d[i] for i, nid in enumerate(node_ids)}
