requirements.txt
streamlit>=1.28.0
networkx>=3.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scipy>=1.11.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
